Break apart template module
